requirements.txt
streamlit
pandas
pdfminer.six
transformers
torch 
langchain
langchain-community
sentence-transformers
faiss-cpu